backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-multipart
pdf2image
